refactor file strucutre by separating the style files to assets folder
