+++ frontend/src/App.tsx
@@ -1,7 +1,5 @@
 import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import CreateSong from "./components/CreateSong";
-import DeleteSong from "./components/DeleteSong";
 import ListSongs from "./components/ListSongs";
 import ShowSongStats from "./components/SongStats/ShowSongStats";
 import { ToastContainer } from 'react-toastify';
