+++ frontend/src/assets/style/ModalStyle.ts
