+++ frontend/src/assets/style/NavbarStyle.ts
