+++ frontend/src/assets/style/ShowSongStatsStyle.ts
