+++ frontend/src/components/common/Navbar/LogoItem.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Logo } from "./NavbarStyle";
+import { Logo } from "../../../assets/style/NavbarStyle";
 
 const LogoItem: React.FC = () => {
   return <Logo>Songs</Logo>;
